Move get_probe_serial to method on Cli
