+++ cmd/gdb/src/lib.rs
@@ -19,8 +19,6 @@
 
 use std::process::{Command, Stdio};
 
-use cmd_openocd::get_probe_serial;
-
 use humility_cli::{ExecutionContext, Subcommand};
 use humility_cmd::{Archive, Command as HumilityCmd, CommandKind};
 
@@ -62,7 +60,7 @@ fn gdb(context: &mut ExecutionContext) -> Result<()> {
     }
 
     let subargs = GdbArgs::try_parse_from(subargs)?;
-    let serial = get_probe_serial(&context.cli, subargs.serial.clone())?;
+    let serial = context.cli.get_probe_serial(subargs.serial.as_deref())?;
 
     let work_dir = tempfile::tempdir()?;
     let name = match &hubris.manifest.name {
